requirement.txt
pillow
line-bot-sdk
flask